# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Softmotriz",
        "description": "rutas de api para consumir",
        "contact": {
            "email": "harveympv@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Softmotriz/public",
            "description": " API Server"
        }
    ],
    "paths": {
        "/api/Schooling": {
            "get": {
                "tags": [
                    "Schooling"
                ],
                "summary": "Mostrar todos los niveles de escolaridad",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "post": {
                "tags": [
                    "Schooling"
                ],
                "summary": "Registrar nivel de escolaridad",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level_schooling"
                                ],
                                "properties": {
                                    "level_schooling": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Primaria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/Schoolings/{Schooling_id}": {
            "get": {
                "tags": [
                    "Schooling"
                ],
                "summary": "Mostrar nivel de escolaridad por ID",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "put": {
                "tags": [
                    "Schooling"
                ],
                "summary": "Actualizar nombre de escolaridad",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level_schooling"
                                ],
                                "properties": {
                                    "level_schooling": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Primaria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Schooling"
                ],
                "summary": "Eliminar nivel de escolaridad",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/BloodType": {
            "get": {
                "tags": [
                    "BloodType"
                ],
                "summary": "Mostrar todos los tipo de sangre",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "post": {
                "tags": [
                    "BloodType"
                ],
                "summary": "Registrar tipo de sangre",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "O+"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/BloodType/{BloodType_id}": {
            "get": {
                "tags": [
                    "BloodType"
                ],
                "summary": "Mostrar tipo de sangre por ID",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "put": {
                "tags": [
                    "BloodType"
                ],
                "summary": "Actualizar tipo de sangre",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "A-"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "delete": {
                "tags": [
                    "BloodType"
                ],
                "summary": "Eliminar tipo de sangre",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/Conventions": {
            "get": {
                "tags": [
                    "Conventions"
                ],
                "summary": "Mostrar todos los convenios",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "post": {
                "tags": [
                    "Conventions"
                ],
                "summary": "Registrar los convenios",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "HACHE"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "12345678"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mf k rosas"
                                    },
                                    "department": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Nariño"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Pasto"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3225785525"
                                    },
                                    "representative": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Kira"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/Conventions/{Conventionsid}": {
            "put": {
                "tags": [
                    "Conventions"
                ],
                "summary": "Actualizar los convenios",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level_schooling"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "HACHE"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "12345678"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mf k rosas"
                                    },
                                    "department": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Nariño"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Pasto"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3225785525"
                                    },
                                    "representative": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Kira"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
...
        "/api/DocumentType": {
            "get": {
                "tags": [
                    "DocumentType"
                ],
                "summary": "Mostrar todos los niveles de escolaridad",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "post": {
                "tags": [
                    "DocumentType"
                ],
                "summary": "Registrar tipo de documento",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "CC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/DocumentType/{DocumentType_id}": {
            "get": {
                "tags": [
                    "DocumentType"
                ],
                "summary": "Mostrar tipo de documento por ID",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "put": {
                "tags": [
                    "DocumentType"
                ],
                "summary": "Actualizar tipo de documento",
                "requestBody": {
                    "description": "dato enviar por formulario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "CC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            },
            "delete": {
                "tags": [
                    "DocumentType"
                ],
                "summary": "Eliminar tipo de documento",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
...
    },
...
}